refactor: extract relationship component
Removes the relationship component imports and logic from the Field
component's switch logic for better readability.
